c/cpp: Use compile_commands if present to determine compiler args
The old behavior without a compile_commands.json present in the working
dir was kept available. If a this file is present, the new mechanism
will use the actual compiler flags as specified by the build system to
parse the actual file the user is working in and retrieve info about the
token under the user's cursor.