Fix neovim unresponsiveness on `doge#install()`
Neovim became unresponsive due to the long-running curl in the
background. But, the user had no indication about that. Now, the
install script is run in a terminal buffer to show the progress
without blocking the editor.

Also included in this commit is a small modification to the command to
run the powershell script. On windows systems, the user does not have
permission to run powershell scripts (by default). The included
modification tries to implement a workaround for that.