Improvements to install.sh
Now, it uses `sh` instead of `bash`, because `bash` may not be avaiable
in all systems, but `sh` is guaranteed (AFAIK) to be avaiable. Also, now
it checks whether curl is available or not. If not avaiable, it shows a
simple and concise error message and informs the user to install `curl`.