Improve the terminal window
Now it takes up only 4 (8 in Windows) rows instead of the default
half-of-current-window. Also, now it utilizes the full width of the
terminal instead of the current window only.

Why 8 rows in Windows ? Well... Invoke-WebRequest shows progress
using a banner that is 4-row-tall, and it also has a minimum top
padding of 2 rows. So, I chose 8 rows to give 2 rows bottom
padding, resulting in a more "balanced" look.